refactor code
refactoring cart.jsx
